Move SchemaMaterializer to its own file
